requirements.txt
streamlit==1.39.0
vtracer==0.6.11
ezdxf==1.3.4
svgpathtools==1.6.1
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
